SCbackend/requirements.txt
asgiref==3.7.2
dj-database-url==2.1.0
Django==4.2.6
django-environ==0.11.2
django-filter==23.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
gunicorn==21.2.0
Markdown==3.4.4
packaging==23.2
psycopg2-binary==2.9.9
PyJWT==2.8.0
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.8.0
